Refactor test folders.
