+++ src/test/unit/java/com/leverx/learningmanagementsystem/utils/TestUtils.java
@@ -1,0 +1,13 @@
+package com.leverx.learningmanagementsystem.utils;
+
+public class TestUtils {
+
+    public static final String CLEAN_SQL = "/sql/clean-db.sql";
+    public static final String INSERT_SQL = "/sql/insert-test-data.sql";
+    public static final String PAGE = "page";
+    public static final String PAGE_SIZE = "page-size";
+    public static final String DEFAULT_PAGE = "0";
+
+    private TestUtils() {
+    }
+}
